+++ src/oidc-gen/gen_handler.h
@@ -6,7 +6,7 @@
 #include "utils/oidc_error.h"
 
 void manualGen(struct oidc_account* account, const struct arguments* arguments);
-void reauthenticate(const char* shortname, const struct arguments* arguments);
+void reauthenticate(const char* shortname, struct arguments* arguments);
 void handleGen(struct oidc_account* account, const struct arguments* arguments,
                const char* cryptPass);
 struct oidc_account* manual_genNewAccount(struct oidc_account*    account,