+++ src/oidc-gen/oidc-gen_options.h
@@ -34,6 +34,7 @@ struct arguments {
   char* pw_cmd;
   char* pw_file;
   char* pw_env;
+  char* pw_gpg;
   char* file;
 
   char* client_id;