+++ src/utils/password_entry.h
@@ -15,6 +15,7 @@ struct password_entry {
   char*         command;
   time_t        expires_after;
   char*         filepath;
+  char*         gpg_key;
 };
 
 #define PW_TYPE_MEM 0x01
@@ -22,6 +23,7 @@ struct password_entry {
 #define PW_TYPE_CMD 0x04
 #define PW_TYPE_PRMT 0x08
 #define PW_TYPE_FILE 0x10
+#define PW_TYPE_GPG 0x20
 
 #define PW_KEY_SHORTNAME IPC_KEY_SHORTNAME
 #define PW_KEY_TYPE "type"
@@ -30,6 +32,7 @@ struct password_entry {
 #define PW_KEY_COMMAND "command"
 #define PW_KEY_EXPIRESAFTER "expires_after"
 #define PW_KEY_PWFILE "pw_file"
+#define PW_KEY_GPG "gpg_key"
 
 void                   _secFreePasswordEntry(struct password_entry*);
 cJSON*                 passwordEntryToJSON(const struct password_entry*);
@@ -41,6 +44,7 @@ void pwe_setCommand(struct password_entry* pw, char* command);
 void pwe_setFile(struct password_entry* pw, char* file);
 void pwe_setShortname(struct password_entry* pw, char* shortname);
 void pwe_setType(struct password_entry* pw, unsigned char type);
+void pwe_setGPGKey(struct password_entry* pw, char* key_id);
 void pwe_setExpiresAt(struct password_entry* pw, time_t expires_at);
 void pwe_setExpiresIn(struct password_entry* pw, time_t expires_in);
 void pwe_setExpiresAfter(struct password_entry* pw, time_t expires_after);